modularized all the backend routes
